+++ components/status/ErrorCatchWidget.vue
@@ -30,6 +30,12 @@
       :value.prop="log_filepath"
       :style="textarea__error_cssprops"
     ></textarea>
+    <div class="div_status-error-catch-next-step-txt">
+      <p>
+        Please restart the computer, unplug and plug back in the <br />
+        power for the Mantarray before attempting to use again
+      </p>
+    </div>
     <div class="div__error-button" :style="error_catch_button_cssprops">
       <ButtonWidget
         :button_widget_width="450"
@@ -39,6 +45,7 @@
         :button_names="['Okay']"
         :enabled_color="'#B7B7B7'"
         :hover_color="['#FFFFFF']"
+        @btn-click="process_ok"
       >
       </ButtonWidget>
     </div>
@@ -46,7 +53,6 @@
 </template>
 <script>
 import ButtonWidget from "@/components/playback/controls/player/ButtonWidget.vue";
-// import { ButtonWidget } from "@/dist/mantarray.common";
 export default {
   name: "ErrorCatchWidget",
   components: {
@@ -57,25 +63,29 @@ export default {
   },
   computed: {
     compute_number_of_rows: function () {
-      return Math.ceil(((this.log_filepath.length * 1.0) / 50).toFixed(1));
+      return Math.ceil(((this.log_filepath.length * 1.0) / 40).toFixed(1));
     },
     error_background_cssprops: function () {
-      return "height: " + (180 + this.compute_number_of_rows * 12) + "px;";
+      return "height: " + (220 + this.compute_number_of_rows * 12) + "px;";
     },
     textarea__error_cssprops: function () {
-      return "height: " + (25 + this.compute_number_of_rows * 10) + "px;";
+      return "height: " + (25 + this.compute_number_of_rows * 12) + "px;";
     },
     error_catch_button_cssprops: function () {
       return (
         "top: " +
-        (160 + this.compute_number_of_rows * 12) +
+        (220 + this.compute_number_of_rows * 12) +
         "px; left: 0px; position: absolute"
       );
     },
   },
   watch: {},
   created() {},
-  methods: {},
+  methods: {
+    process_ok: function (choice) {
+      this.$emit("ok-clicked");
+    },
+  },
 };
 </script>
 <style>
@@ -178,6 +188,32 @@ a:hover {
   font-style: normal;
   text-decoration: none;
   resize: none;
+  z-index: 5;
+  pointer-events: all;
+}
+
+.div_status-error-catch-next-step-txt {
+  line-height: 1.2;
+  transform: rotate(0deg);
+  padding: 0px;
+  margin: 0px;
+  overflow-wrap: break-word;
+  color: rgb(183, 183, 183);
+  font-family: Muli;
+  position: absolute;
+  top: 190px;
+  left: 0px;
+  width: 450px;
+  height: 66px;
+  overflow: hidden;
+  visibility: visible;
+  user-select: none;
+  text-align: center;
+  font-size: 15px;
+  letter-spacing: normal;
+  font-weight: normal;
+  font-style: normal;
+  text-decoration: none;
   z-index: 5;
   pointer-events: all;
 }