+++ components/status/StatusBar.vue
@@ -1,12 +1,36 @@
 <template>
   <div class="div__status-bar">
     <span class="span__status-bar-text">{{ alert_txt }}</span>
+    <span>
+      <b-modal
+        id="error-catch"
+        size="sm"
+        hide-footer
+        hide-header
+        hide-header-close
+        :static="true"
+      >
+        <ErrorCatchWidget
+          id="error"
+          :log_filepath="log_path"
+          @ok-clicked="remove_error_catch"
+        ></ErrorCatchWidget>
+      </b-modal>
+    </span>
   </div>
 </template>
 <script>
-import { mapGetters } from "vuex";
+import Vue from "vue";
+import { mapGetters, mapState } from "vuex";
 import { STATUS } from "@/store/modules/flask/enums";
+import BootstrapVue from "bootstrap-vue";
+import { BButton } from "bootstrap-vue";
+import { BModal } from "bootstrap-vue";
+import ErrorCatchWidget from "@/components/status/ErrorCatchWidget.vue";
 
+Vue.use(BootstrapVue);
+Vue.component("BButton", BButton);
+Vue.component("BModal", BModal);
 /**
  * @vue-data     {String} alert_txt - Contains the current status of the Application and its updated as status change.
  * @vue-computed {String} status_uuid - Contains a UUID which represents a meaningful information, from Vuex store.
@@ -14,6 +38,9 @@ import { STATUS } from "@/store/modules/flask/enums";
  */
 export default {
   name: "StatusBar",
+  components: {
+    ErrorCatchWidget,
+  },
   data() {
     return {
       alert_txt: "",
@@ -23,13 +50,15 @@ export default {
     ...mapGetters({
       status_uuid: "flask/status_id",
     }),
+    ...mapState("settings", {
+      log_path: "log_path",
+    }),
   },
   watch: {
     status_uuid: function (newValue) {
       this.set_text_from_state(newValue);
     },
   },
-
   created() {
     this.set_text_from_state(this.status_uuid);
   },
@@ -65,11 +94,28 @@ export default {
           this.alert_txt += `Recording to File`;
           break;
         case STATUS.MESSAGE.ERROR:
+          this.shutdown_request();
           this.alert_txt += `Error Occurred`;
+          this.$bvModal.show("error-catch");
+          break;
+        case STATUS.MESSAGE.SHUTDOWN:
+          this.alert_txt += `Shutting Down`;
           break;
         default:
           this.alert_txt = `Status:` + new_value; // to be 43 characters and include the UUID, there isn't room for a space
           break;
+      }
+    },
+    remove_error_catch: function () {
+      this.$bvModal.hide("error-catch");
+      this.$store.commit("flask/set_status_uuid", STATUS.MESSAGE.SHUTDOWN);
+    },
+    shutdown_request: async function () {
+      const shutdown_url = "http://localhost:4567/shutdown";
+      try {
+        await Vue.axios.get(shutdown_url);
+      } catch (error) {
+        return;
       }
     },
   },