+++ tests/unit/store/flask.spec.js
@@ -291,7 +291,14 @@ describe("store/flask", () => {
           STATUS.MESSAGE.ERROR
         );
       });
+      test("Given the status is set to ERROR, When attempting to commit SHUTDOWN, Then updates to SHUTDOWN mode", () => {
+        store.commit("flask/set_status_uuid", STATUS.MESSAGE.ERROR);
 
+        store.commit("flask/set_status_uuid", STATUS.MESSAGE.SHUTDOWN);
+        expect(store.state.flask.status_uuid).toStrictEqual(
+          STATUS.MESSAGE.SHUTDOWN
+        );
+      });
       test("When set_status_ping_interval_id is committed, the ID mutates", async () => {
         const expected_id = 2993;
         store.commit("flask/set_status_ping_interval_id", expected_id);